feat: Move getCanvasContext function to a separate file
